fix PR comments to make ProactiveStateMiddleware Generic
