+++ sdk/csharp/libraries/microsoft.bot.solutions/Proactive/Sharded/ProactiveStateMiddlewareSharded.cs
@@ -13,44 +13,11 @@ namespace Microsoft.Bot.Solutions.Proactive.Sharded
     /// A Middleware for saving the proactive model data
     /// This middleware will refresh user's latest conversation reference and save it to state.
     /// </summary>
-    public class ProactiveStateMiddlewareSharded : IMiddleware
+    public class ProactiveStateMiddlewareSharded : ProactiveStateMiddleWareTemplate<ProactiveStateSharded>
     {
-        private ProactiveStateSharded _proactiveState;
-        private IStatePropertyAccessor<ProactiveModel> _proactiveStateAccessor;
-
-        public ProactiveStateMiddlewareSharded(ProactiveStateSharded proactiveState)
+        public ProactiveStateMiddlewareSharded(ProactiveStateSharded proactiveStateSharded)
+            : base(proactiveStateSharded)
         {
-            _proactiveState = proactiveState;
-            _proactiveStateAccessor = _proactiveState.CreateProperty<ProactiveModel>(nameof(ProactiveModel));
-        }
-
-        public async Task OnTurnAsync(ITurnContext turnContext, NextDelegate next, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            var activity = turnContext.Activity;
-
-            if (!string.IsNullOrEmpty(activity.From.Role) && activity.From.Role.Equals("user", StringComparison.InvariantCultureIgnoreCase))
-            {
-                var proactiveState = await _proactiveStateAccessor.GetAsync(turnContext, () => new ProactiveModel()).ConfigureAwait(false);
-                ProactiveModel.ProactiveData data;
-                var hashedUserId = MD5Util.ComputeHash(turnContext.Activity.From.Id);
-                var conversationReference = turnContext.Activity.GetConversationReference();
-                var proactiveData = new ProactiveModel.ProactiveData { Conversation = conversationReference };
-
-                if (proactiveState.TryGetValue(hashedUserId, out data))
-                {
-                    data.Conversation = conversationReference;
-                }
-                else
-                {
-                    data = new ProactiveModel.ProactiveData { Conversation = conversationReference };
-                }
-
-                proactiveState[hashedUserId] = data;
-                await _proactiveStateAccessor.SetAsync(turnContext, proactiveState).ConfigureAwait(false);
-                await _proactiveState.SaveChangesAsync(turnContext).ConfigureAwait(false);
-            }
-
-            await next(cancellationToken).ConfigureAwait(false);
         }
     }
 }
